Separate class in MainWindows.cs
