+++ TCObjectStorageClient/MainWindow.xaml.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Input;
 using TCObjectStorageClient.ViewModels;
 
 namespace TCObjectStorageClient
@@ -29,54 +27,6 @@ namespace TCObjectStorageClient
                     vm.Password = box.Password;
                 }
             }
-        }
-    }
-
-    public abstract class CommandBase : ICommand
-    {
-        protected readonly MainViewModel _viewModel;
-
-        public CommandBase(MainViewModel viewModel)
-        {
-            _viewModel = viewModel;
-        }
-
-        public bool CanExecute(object parameter) => true;
-
-        public void Execute(object parameter)
-        {
-            ExecuteInternal();
-        }
-
-        public event EventHandler CanExecuteChanged;
-
-        protected abstract void ExecuteInternal();
-    }
-
-    public class DelegateCommand : CommandBase
-    {
-        private readonly Action _action;
-
-        public DelegateCommand(MainViewModel viewModel, Action action) : base(viewModel)
-        {
-            _action = action;
-        }
-
-        protected override void ExecuteInternal()
-        {
-            _action();
-        }
-    }
-
-    namespace ViewModels
-    {
-        public partial class MainViewModel
-        {
-            public ICommand OpenFileCommand => new DelegateCommand(this, UploadFiles);
-
-            public ICommand OpenDirectoryCommand => new DelegateCommand(this, UploadDirectory);
-
-            public ICommand DeleteFilesCommand => new DelegateCommand(this, DeleteAllFilesInContainer);
         }
     }
 }
