+++ TCObjectStorageClient/TCObjectStorageClient.WPF.csproj
@@ -61,6 +61,9 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="Commands\CommandBase.cs" />
+    <Compile Include="Commands\DelegateCommand.cs" />
+    <Compile Include="MainViewModel.Commands.cs" />
     <Compile Include="Properties\Annotations.cs" />
     <Compile Include="WindowsAlertDialog.cs" />
     <Compile Include="WindowsFileImporter.cs" />
